server/train_ai_model.py
import os
import sys
import subprocess
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

import logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

def fetch_latest_dataset():
    # Implement dataset fetching logic
    logger.info("Fetching latest dataset...")
    # Example: git pull or download from cloud storage
    subprocess.run(["git", "pull"], cwd="data")
    logger.info("Dataset fetched.")

def train_model():
    # Implement AI model training logic using a cloud service
    logger.info("Training AI model...")
    # Example: Using AWS SageMaker CLI
    # Ensure AWS credentials are configured
    subprocess.run([
        "aws", "sagemaker", "create-training-job",
        "--training-job-name", "think-ai-training-job",
        "--algorithm-specification", "TrainingImage=382416733822.dkr.ecr.us-west-2.amazonaws.com/tensorflow-training:2.3.0,TrainingInputMode=File",
        "--role-arn", os.getenv("SAGEMAKER_ROLE_ARN"),
        "--input-data-config", "file://training_input.json",
        "--output-data-config", "S3OutputPath=s3://your-bucket/think-ai-output/",
        "--resource-config", "InstanceType=ml.m5.large,InstanceCount=1,VolumeSizeInGB=10",
        "--stopping-condition", "MaxRuntimeInSeconds=86400"
    ])
    logger.info("Model training initiated.")

def evaluate_model():
    # Implement model evaluation logic
    logger.info("Evaluating AI model...")
    # Example: Run evaluation script
    result = subprocess.run([sys.executable, "evaluate_model.py"], capture_output=True, text=True)
    if result.returncode != 0:
        logger.error("Model evaluation failed.")
        return 0.0
    accuracy = float(result.stdout.strip())
    logger.info(f"Model accuracy: {accuracy}%")
    return accuracy

def deploy_model():
    # Implement model deployment logic
    logger.info("Deploying AI model...")
    # Example: Using AWS SageMaker
    subprocess.run([
        "aws", "sagemaker", "create-model",
        "--model-name", "think-ai-model",
        "--primary-container", "Image=382416733822.dkr.ecr.us-west-2.amazonaws.com/tensorflow-inference:2.3.0,ModelDataUrl=s3://your-bucket/think-ai-output/model.tar.gz",
        "--execution-role-arn", os.getenv("SAGEMAKER_ROLE_ARN")
    ])
    logger.info("Model deployed.")

def rollback_model():
    # Implement AI model rollback logic
    logger.info("Rolling back AI model due to performance degradation...")
    # Example: Redeploy previous stable model
    subprocess.run([
        "aws", "sagemaker", "update-model",
        "--model-name", "think-ai-model",
        "--primary-container", "Image=382416733822.dkr.ecr.us-west-2.amazonaws.com/tensorflow-inference:2.2.0,ModelDataUrl=s3://your-bucket/think-ai-output/previous_model.tar.gz",
        "--execution-role-arn", os.getenv("SAGEMAKER_ROLE_ARN")
    ])
    logger.info("Model rollback completed.")

def main():
    try:
        fetch_latest_dataset()
        train_model()
        accuracy = evaluate_model()
        if accuracy >= 98.0:
            deploy_model()
        else:
            logger.warning("Model accuracy below threshold. Initiating rollback.")
            rollback_model()
    except Exception as e:
        logger.error(f"An error occurred: {e}")
        sys.exit(1)

if __name__ == "__main__":
    main()